pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gtfs-archive-txt"
version = "0.1.0"
description = "GTFS archive text application"
readme = "app/README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "TransitMatters", email = "team@transitmatters.org"}
]
dependencies = [
    "requests==2.31.0",
    "python-dotenv==1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
# UV configuration

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ['py38']

[dependency-groups]
dev = [
    "ruff>=0.14.6",
]
